ingest: Split test_utils out from script_utils
